+++ src/commands/lint.rs
@@ -66,17 +66,36 @@ fn validate_single(
     let (single_read_validators, _) =
         validators::filter_validators(single_read_validation_level, None, disabled_validators);
 
-    let span = info_span!("validate_single");
-    let _span_ctx = span.enter();
+    let mut duplicate_name_validator = DuplicateNameValidator::new();
+
+    let code = duplicate_name_validator.code();
+    let name = duplicate_name_validator.name();
+    let use_special_validator = !disabled_validators.contains(&code.to_string());
+
+    let validators = if use_special_validator {
+        format!(r#""[{code}] {name}""#)
+    } else {
+        String::new()
+    };
+
+    info!("enabled special validators: [{}]", validators);
+
+    let span = info_span!("validate_single", pass = 1);
+    let span_ctx = span.enter();
 
     info!("start");
 
     let mut record = Record::default();
+
     let mut record_counter = 0;
     let mut failure_count = 0;
 
     while reader.read_record(&mut record)? != 0 {
         record.reset(record_definition_separator);
+
+        if use_special_validator {
+            duplicate_name_validator.insert(&record);
+        }
 
         for validator in &single_read_validators {
             if let Err(e) = validator.validate(&record) {
@@ -87,6 +106,27 @@ fn validate_single(
 
         record_counter += 1;
     }
+
+    info!(record_count = record_counter, "end");
+    drop(span_ctx);
+
+    if !use_special_validator {
+        return Ok(failure_count);
+    }
+
+    let span = info_span!("validate_single", pass = 2);
+    let _span_ctx = span.enter();
+
+    info!("start");
+
+    let (record_counter, duplicate_name_failure_count) = validate_duplicate_names(
+        r1_src,
+        record_definition_separator,
+        lint_mode,
+        duplicate_name_validator,
+    )?;
+
+    failure_count += duplicate_name_failure_count;
 
     info!(record_count = record_counter, "end");